SOP-33 Manage "Users-orgstructure" relation

+++ app/services/UserService.scala
@@ -5,11 +5,12 @@ import javax.inject.{Inject, Singleton}
 import com.mohiva.play.silhouette.api.LoginInfo
 import com.mohiva.play.silhouette.api.services.IdentityService
 import com.mohiva.play.silhouette.impl.providers.CommonSocialProfile
-import models.dao.UserDao
+import models.dao.{UserDao, UserGroupDao}
 import models.user.{User => UserModel}
+import org.davidbild.tristate.Tristate
 import play.api.libs.concurrent.Execution.Implicits._
 import services.authorization.UserSda
-import utils.errors.NotFoundError
+import utils.errors.{ConflictError, NotFoundError}
 import utils.listmeta.ListMeta
 
 import scala.async.Async.{async, await}
@@ -20,7 +21,8 @@ import scala.concurrent.Future
   */
 @Singleton
 class UserService @Inject()(
-  protected val userDao: UserDao
+  protected val userDao: UserDao,
+  protected val userGroupDao: UserGroupDao
 ) extends IdentityService[UserModel]
   with ServiceResults[UserModel] {
 
@@ -73,15 +75,19 @@ class UserService @Inject()(
     * @param role    role filter
     * @param status  status filter
     * @param account logged in user
+    * @param groupId only users of the group
     * @param meta    list meta
     */
   def list(
     role: Option[UserModel.Role],
-    status: Option[UserModel.Status]
+    status: Option[UserModel.Status],
+    groupId: Tristate[Long]
   )(implicit account: UserModel, meta: ListMeta): ListResult = async {
     val users = await(userDao.getList(
+      id = None,
       role = role,
-      status = status
+      status = status,
+      groupId = groupId
     ))
     users
   }
@@ -118,8 +124,12 @@ class UserService @Inject()(
     await(getById(id)) match {
       case Left(error) => error
       case Right(original) =>
-        await(userDao.delete(id))
-        unitResult
+        if (await(userGroupDao.exists(userId = Some(id)))) {
+          ConflictError.User.GroupExists(id)
+        } else {
+          await(userDao.delete(id))
+          unitResult
+        }
     }
   }
 }
